refactor(kubernetes): deprecate ProviderVersion
* Since ProviderVersion was only ever relevant to KubernetesAccount, move the enum from Provider to KubernetesAccount before deprecating it. Hopefully this will prevent any future cloud providers from needlessly including the providerVersion field in every account.
* We have included warnings that V1 accounts cannot be used for a Distributed install for a couple of Halyard releases; let's remove these for the next release.
