+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/model/v1/node/Provider.java
@@ -72,22 +72,6 @@ public abstract class Provider<A extends Account> extends Node implements Clonea
 
   public abstract ProviderType providerType();
 
-  public enum ProviderVersion {
-    V1("v1"),
-    V2("v2");
-
-    private final String name;
-
-    ProviderVersion(String name) {
-      this.name = name;
-    }
-
-    @Override
-    public String toString() {
-      return this.name;
-    }
-  }
-
   public enum ProviderType {
     APPENGINE("appengine"),
     AWS("aws"),
